Split MainWindow logic into partial classes
