+++ PSSG Editor/AttributeItem.cs
@@ -1,0 +1,10 @@
+namespace PSSGEditor
+{
+    internal class AttributeItem
+    {
+        public string Key { get; set; }
+        public string Value { get; set; }
+        public int OriginalLength { get; set; }
+    }
+}
+
